# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AlloyDB API",
    "version": "1.0.0",
    "description": "API for managing Users, Items, Orders, and Ratings in AlloyDB"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of users" }
        }
      },
      "post": {
        "summary": "Create a user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "User details" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated" }
        }
      },
      "delete": {
        "summary": "Delete user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "User deleted" }
        }
      }
    },
    "/users/{id}/orders": {
      "get": {
        "summary": "Get user's order history",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "User order history with items" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/items": {
      "get": {
        "summary": "Get all items",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of items" }
        }
      },
      "post": {
        "summary": "Create an item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item_description": { "type": "string" },
                  "item_value": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Item created" }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "summary": "Get item by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Item details" },
          "404": { "description": "Item not found" }
        }
      },
      "put": {
        "summary": "Update item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item_description": { "type": "string" },
                  "item_value": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Item updated" }
        }
      },
      "delete": {
        "summary": "Delete item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Item deleted" }
        }
      }
    },
    "/items/{id}/average-rating": {
      "get": {
        "summary": "Get average rating for an item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Average rating" }
        }
      }
    },
    "/items/{id}/top-complaints": {
      "get": {
        "summary": "Get top complaints analysis for an item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "AI-generated top complaints summary" },
          "404": { "description": "Item or comments not found" }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get all orders",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of orders" }
        }
      },
      "post": {
        "summary": "Create an order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "create_date": { "type": "string", "format": "date" },
                  "status": { "type": "string" },
                  "user_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created" }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "404": { "description": "Order not found" }
        }
      },
      "put": {
        "summary": "Update order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "create_date": { "type": "string", "format": "date" },
                  "status": { "type": "string" },
                  "user_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order updated" }
        }
      },
      "delete": {
        "summary": "Delete order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Order deleted" }
        }
      }
    },
    "/order_items": {
      "get": {
        "summary": "Get all order items",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of order items" }
        }
      },
      "post": {
        "summary": "Create an order item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": { "type": "integer" },
                  "item_id": { "type": "integer" },
                  "quantity": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order item created" }
        }
      }
    },
    "/order_items/{id}": {
      "get": {
        "summary": "Get order item by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Order item details" },
          "404": { "description": "Order item not found" }
        }
      },
      "put": {
        "summary": "Update order item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": { "type": "integer" },
                  "item_id": { "type": "integer" },
                  "quantity": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order item updated" }
        }
      },
      "delete": {
        "summary": "Delete order item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Order item deleted" }
        }
      }
    },
    "/ratings": {
      "get": {
        "summary": "Get all ratings",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of ratings" }
        }
      },
      "post": {
        "summary": "Create a rating",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": { "type": "integer" },
                  "comments": { "type": "string" },
                  "user_id": { "type": "integer" },
                  "order_items_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Rating created" }
        }
      }
    },
    "/ratings/{id}": {
      "get": {
        "summary": "Get rating by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Rating details" },
          "404": { "description": "Rating not found" }
        }
      },
      "put": {
        "summary": "Update rating",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": { "type": "integer" },
                  "comments": { "type": "string" },
                  "user_id": { "type": "integer" },
                  "order_items_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Rating updated" }
        }
      },
      "delete": {
        "summary": "Delete rating",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Rating deleted" }
        }
      }
    }
  }
}